+++ packages/common/src/index.ts
@@ -1,6 +1,7 @@
-export * from './utils';
-export * from './data-sorting';
-export * from './types';
 export * from './config';
+export * from './data-sorting';
 export * from './domain/feature-flags';
 export * from './domain/lokalize';
+export * from './intl';
+export * from './types';
+export * from './utils';
