+++ packages/e2e/package.json
@@ -4,10 +4,10 @@
   "version": "0.0.0",
   "license": "EUROPEAN UNION PUBLIC LICENCE v. 1.2",
   "devDependencies": {
-    "@types/node": "^14.14.25",
+    "@types/node": "^14.14.41",
     "@typescript-eslint/eslint-plugin": "^4.14.2",
     "@typescript-eslint/parser": "^4.14.2",
-    "cypress": "^6.4.0",
+    "cypress": "^7.1.0",
     "eslint": "^7.19.0",
     "eslint-config-prettier": "^7.1.0",
     "eslint-plugin-prettier": "^3.1.4",
