+++ packages/e2e/src/integration/gemeente/positief-geteste-mensen.spec.ts
@@ -1,19 +1,25 @@
-// import { MunicipalContext } from '~/integration/types';
+import { MunicipalContext } from '~/integration/types';
 
-// context('Gemeente - Positief geteste mensen', () => {
-//   before(() => {
-//     cy.beforeMunicipalTests('positief-geteste-mensen');
-//   });
+context('Gemeente - Positief geteste mensen', () => {
+  before(() => {
+    cy.beforeMunicipalTests('positief-geteste-mensen');
+  });
 
-//   // @ts-ignore
-//   it('Should show the correct KPI values', function (this: MunicipalContext) {
-//     const lastValue = this.municipalData.tested_overall.last_value;
+  // @ts-ignore
+  it('Should show the correct KPI values', function (this: MunicipalContext) {
+    const lastValue = this.municipalData.tested_overall.last_value;
 
-//     const kpiTestInfo = {
-//       infected_per_100k: formatNumber(lastValue.infected_per_100k),
-//       infected: formatNumber(lastValue.infected),
-//     };
+    console.log(cy.formatters.formatNumber(100));
 
-//     cy.checkKpiValues(kpiTestInfo);
-//   });
-// });
+    const kpiTestInfo = {
+      infected_per_100k: cy.formatters.formatNumber(
+        lastValue.infected_per_100k
+      ),
+      infected: cy.formatters.formatNumber(lastValue.infected),
+    };
+
+    console.dir(kpiTestInfo);
+
+    cy.checkKpiValues(kpiTestInfo);
+  });
+});
