+++ packages/e2e/src/support/index.ts
@@ -1,1 +1,2 @@
 import './commands';
+import './formatters';
